app/requirements.txt
smbus
pimoroni-bme280
bmp280
picamera
pillow
Jinja2
gpiozero